feat: added command input file
